refactor: apis/team 폴더구조 변경
